% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnet.R
\name{accnet}
\alias{accnet}
\title{Build accessory genome network (AccNET)}
\usage{
accnet(mmseqs, threshold = 0.8, singles = TRUE)
}
\arguments{
\item{mmseqs}{An object od class \emph{mmseq}}

\item{threshold}{Remove proteins with a frequency higher than threshold (soft-coregenome).}

\item{singles}{keep proteins presence in just one sample?}
}
\value{
An accnet object.\cr
\emph{accnet} object contains\cr
\itemize{
\item{\emph{list}}: Data.frame with the relationship of genome-protein
\item{\emph{matrix}}: Adjacency matrix (presence/absence)
\item{\emph{annot}}: Annotation table with the protein functional annotation
}
}
\description{
Create an \emph{accnet} object.
}
\details{
This function extract the accessory genome from a \emph{mmseq} object.
The \emph{mmseq} contains all the relationship among genomes (or other
genomic unit) and genes/proteins. \emph{accnet} extract the accsessory
genome removing the sof-coregenome. Hard-coregenome is the set of
proteins present in the 100\\% of the genomes. Soft-coregenome is the set
of proteins present in the \emph{threshold} of the genomes.
}
\references{
AcCNET (Accessory Genome Constellation Network): comparative genomics software for accessory genome analysis using bipartite networks
VF Lanza, F Baquero, F de la Cruz, TM Coque - Bioinformatics, 2017
}
\seealso{
\emph{\link{mmseqs}}
}
