% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnet_enrichment_analysis.R
\name{accnet_enrichment_analysis}
\alias{accnet_enrichment_analysis}
\title{Enrichment analysis of accessory genome}
\usage{
accnet_enrichment_analysis(data, cluster, padj_method = "BY")
}
\arguments{
\item{data}{An \code{accnet} object}

\item{cluster}{Dataframe with two colums (Source,Cluster)}

\item{padj_method}{Method adjust the p-value for multiple comparisons:
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none" (see p.adjust)}

\item{max_pvalue}{Maximun p-value of retrived data.}
}
\value{
Return a data.frame with the columns:
\itemize{
\item Target: Protein
\item Source: Genome
\item Cluster: Cluster of genome
\item perClusterFreq: Percentage (\\%) of the protein in this cluster
\item ClusterFreq: Frequency \link{0-1} of the protein in this cluster
\item ClusterGenomeSize: Number of genomes in the cluster
\item perTotalFreq: Percentage (\\%) of the protein in the population
\item TotalFreq: Frequency \link{0-1} of the protein in the population
\item OdsRatio: Ods ration of the protein
\item pvalue: p-value of the hypergeometric test (see \code{phyper})
\item padj: Adjusted p-value using \code{p.adjust}
\item AccnetGenomeSize: Total number of genomes
\item AccnetProteinSize: Total number of proteins
\item Annot: Funtional annotation of the protein.
}
}
\description{
Perform a statistical test (Hypergeometric test: \code{phyper}) of the genes/proteins.
It takes the cluster definition from the \code{cluster} table and compute
the statistical test assuming the frequency of the gene/protein in the cluster vs
the frequency of the gene/protein in the population.
}
\seealso{
\code{\link[stats]{phyper}}

\code{\link[stats]{p.adjust}}

\code{\link{accnet}}
}
