% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnet_with_padj.R
\name{accnet_with_padj}
\alias{accnet_with_padj}
\title{Creates a AccNET network using the enrichment analysis}
\usage{
accnet_with_padj(data)
}
\arguments{
\item{data}{\code{accnet_enrichment_analysis} result.}
}
\value{
Return and \code{accnet} object with the adjusted p-value as edge-weigth
}
\description{
Creates and improved network using the adjusted p-value as edge-weight.
The network could be export to gephi (\code{export_to_gephi}) for a correct
visualitation
}
\seealso{
\code{\link{accnet_enrichment_analysis}}, \code{\link{export_to_gephi}}
}
