% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate Virulence Factors and Antibiotic Resistance Genes}
\usage{
annotate(
  files,
  re_use = TRUE,
  type = "nucl",
  database = c("AbR", "VF_A", "VF_B"),
  query = "all"
)
}
\arguments{
\item{files}{data.frame with the absolute path to the genome files (protein fasta file).}

\item{re_use}{Re-use the precomputed mmseqs2 database. The index must be in
the working directory}

\item{type}{user must be specified if the data set is nucleotide or protein.}

\item{database}{A vector with the query databases:
\itemize{
\item \emph{AbR}: Antibiotic resistance database (ResFinder)
\item \emph{VF_A} VFDB core dataset (genes associated with experimentally verified VFs only)
\item \emph{VF_B} VFDB full dataset (all genes related to known and predicted VFs)
}}

\item{query}{"all" or "accessory". It perform the annotation from whole
protein dataset or just from the accessory}
}
\value{
A \code{data.frame} with the annotation information\cr
\itemize{
\item \emph{Genome}: Genome query
\item \emph{Protein}: Proteins query
\item \emph{target}: Protein subject (AbR o VF)
\item \emph{pident}: Percentage of identical matches
\item \emph{alnlen}: Alingment length
\item \emph{mismatch}: number of mismatchs
\item \emph{gapopen}: number of gaps
\item \emph{qstart}: query start alingment
\item \emph{qend}: query end alingment
\item \emph{tstart}: target start alingment
\item \emph{tend}:  target end alingment
\item \emph{evalue}: evalue
\item \emph{bits}: Bitscore
\item \emph{DataBase}: Database (AbR, VF_A or VF_B)
\item \emph{Gene}: Gene name
\item \emph{Description}: Functional annotation (VF) or category (AbR)
}
}
\description{
This function annotate virulence factors and the antibiotic resitance genes
of the genomes (\code{files}). The annotation is performed using
\strong{mmseqs2} software (\link{https://github.com/soedinglab/MMseqs2}) and the
databases \strong{Virulence Factor DataBase}
(\link{http://www.mgc.ac.cn/cgi-bin/VFs/v5/main.cgi}) and \strong{ResFinder}
(\link{https://cge.cbs.dtu.dk/services/ResFinder/}). The function can re-use the
previous computational steps of \code{mmseqs} or create a new index database from
the files. Re-use option shorten the computational time. This method use the algorithm
\strong{search} of \strong{mmseqs2} so it olny return high identity matchs.
}
\note{
Keep in mind that the results from \emph{accesory} are based on the
annotation of the representative protein of the homologous cluster and therefore
does not mean that all the genomes have the same allele of the gene.
}
