% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.nr_list.R
\name{as.data.frame.nr_list}
\alias{as.data.frame.nr_list}
\title{Wrapper to change format between nr_list and data.frame}
\usage{
as.data.frame.nr_list(data)
}
\arguments{
\item{data}{A \code{nr_list} object}
}
\value{
A \code{data.frame}
}
\description{
Wrapper to change format between nr_list and data.frame
}
