% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_accnet.R
\name{cluster_accnet}
\alias{cluster_accnet}
\title{Internal function to cluster accnet data}
\usage{
cluster_accnet(data, method, n_cluster, d_reduction, ...)
}
\arguments{
\item{data}{\emph{accnet} object}

\item{method}{Clustering method}

\item{n_cluster}{number of cluster (if applicable)}

\item{d_reduction}{boolean. Dimensional reduction using umap}

\item{...}{additional parameters}
}
\value{
\emph{data.frame} with two columns: \emph{Source} and \emph{Target}
}
\description{
Internal function to cluster accnet data
}
