% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_files_from_distance.R
\name{cluster_files_from_distance}
\alias{cluster_files_from_distance}
\title{Internal Function of the workflow pangenomes_from_file}
\usage{
cluster_files_from_distance(files, file_type, distance, n_cores, folder)
}
\arguments{
\item{files}{data.frame with the path of the selected genomes}

\item{file_type}{Type of fasta files 'prot' or 'nucl'}

\item{distance}{threshold distances of the clustering proccess}

\item{n_cores}{Num of cores to use with MASH}

\item{folder}{}
}
\value{
A data frame with the membership of each genome.
}
\description{
Internal Function of the workflow pangenomes_from_file
}
