% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mash.R
\name{cluster_mash}
\alias{cluster_mash}
\title{Internal function to cluster mash data}
\usage{
cluster_mash(data, method, n_cluster, d_reduction, ...)
}
\arguments{
\item{data}{\emph{mash} object}

\item{method}{Clustering method}

\item{n_cluster}{number of cluster (if applicable)}

\item{d_reduction}{boolean. Dimensional reduction using umap}

\item{...}{additional parameters}
}
\value{
\emph{data.frame} with two columns: \emph{Source} and \emph{Target}
}
\description{
Internal function to cluster mash data
}
