% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coincidents.R
\name{coincidents}
\alias{coincidents}
\title{Find coincidents gene/proteins (clustering)}
\usage{
coincidents(data, min_freq, min_size)
}
\arguments{
\item{data}{An \emph{accnet} object}

\item{min_freq}{Minimun frequency of gene/protein in the population.}

\item{min_size}{Minimun size of the reported clusters.}
}
\value{
A data.frame with the protein/gene ID ("Target") and the membership (cluster number)
}
\description{
This function find groups of coincidents genes/proteins in an \emph{accnet}
object. Unlike \emph{accnet_enrichment_analysis()} \emph{concidents} does not need
to define genome/pangenome clusters. The methos perform a multi-dimensional scaling
using \emph{umap} and then find clusters with \emph{hdbscan}.
}
