% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_genome.R
\name{core_genome}
\alias{core_genome}
\title{Core-Genome Alignment}
\usage{
core_genome(data, type, n_cores, method = "fast")
}
\arguments{
\item{data}{An \emph{mmseqs} object}

\item{type}{Type of sequence 'nucl' or 'prot'}

\item{n_cores}{Number of computer core to use}

\item{methos}{\emph{fast (based on blast)} or \emph{accurate (based on mafft)}}
}
\value{
A core_genome object (a data.frame with two columns: fasta header and sequence)
}
\description{
Find and creates a core-genome alignment. Unlike \emph{core_plots()}
this function find the hard core-genome (genes presence in 100\% of genomes
and without repetitions (i.e. without paralougs)). The function takes a
\emph{mmseqs()} output, so the definition of the orthologous genes of the
core-genome (similarity, coverage and/or e-value)depends on the
\emph{mmseqs()} parameters.
}
\details{
The function can performs a pseudo-msa per each ortholog using the function
\emph{result2msa} of \emph{mmseqs2}.This approach is much faster than
classical MSA (clutal, mafft or muscle) but is less accurate. Taking into account
that most of the phylogenetic inference software only takes variant columns with
no insertions or deletion, there are not to many difference in the final phylogenetic trees.

However, \emph{core_genome} also implements an accurate method that use \emph{mafft}
to build a MSA of each gene cluster.

\emph{core_genome()} can build a core-genome alignment of thousands of genomes in minutes.
}
