% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_snps_matrix.R
\name{core_snps_matrix}
\alias{core_snps_matrix}
\title{Core SNPs matrix}
\usage{
core_snps_matrix(data, norm = T)
}
\arguments{
\item{data}{A core_genome object.}

\item{norm}{Number of SNPs normalize by the length of the alingment?}
}
\value{
A square matrix
}
\description{
This function calculates the number of SNPs among samples in core genome.
}
