% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnogmapper.R
\name{eggnogmapper}
\alias{eggnogmapper}
\title{EggNogg Mapper}
\usage{
eggnogmapper(data, n_cores, executable_path = "emapper.py")
}
\arguments{
\item{data}{}
}
\description{
this is a wraper function to external software EggnoggMapper
}
