% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_accnet_aln.R
\name{export_accnet_aln}
\alias{export_accnet_aln}
\title{Export accesory genome binary multi-alignment.}
\usage{
export_accnet_aln(accnet, file, min_freq)
}
\arguments{
\item{accnet}{An \emph{accnet} object}

\item{file}{Output filename}

\item{min_freq}{Minimun frequency of a gene/proteins to include in the alignment}
}
\description{
This function write the presence/abscence matrix into a binary alignment file
that can be use with RaxML or iqtree to build a phylogenetic tree.
}
\details{
Some phylogenetic tree inference tool allow as input binary aligments. We recommend
to use RAxML-NG or iQtree to inference an accessory genome tree.
}
\examples{

}
