% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_core_to_fasta.R
\name{export_core_to_fasta}
\alias{export_core_to_fasta}
\title{Export core-genome alignment to Multi-alignment FASTA file.}
\usage{
export_core_to_fasta(core_data, file)
}
\arguments{
\item{core_data}{A core_genome object}

\item{file}{Path for alignment ouput file}
}
\description{
Export core-genome alignment to Multi-alignment FASTA file.
}
