% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_gephi.R
\name{export_to_gephi}
\alias{export_to_gephi}
\title{Export accnet or igraph object to GraphML format}
\usage{
export_to_gephi(data, file, cluster)
}
\arguments{
\item{data}{\emph{accnet} or \emph{igraph} object}

\item{file}{Filename for network}
}
\description{
This function export a network (\emph{accnet} or \emph{igraph}) to universal
format graphml.
}
\note{
Big networks (especially accnet networks) can not be visualized in R
enviroment. We recomend to use external software such as Gephi or Cytoscape.
Gephi can manage bigger networks and the ForceAtlas2 layout can arrange
accnet networks up to 1.000 genomes. If \emph{accnet} object comes from
\emph{accnet_with_padj()} function, the graph files will include the
edges-weight proportional to the p-value that associate the protein with
the genome.
}
