% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_non_redundant.R
\name{extract_non_redundant}
\alias{extract_non_redundant}
\title{Create a accnet or mash object with the representatives members of each cluster}
\usage{
extract_non_redundant(data, nr_list)
}
\arguments{
\item{data}{\emph{accnet} or \emph{mash} object}

\item{nr_list}{\emph{data.frame} with two columns (Source, Target) or
\emph{nr_list} object created with \emph{non_redundant()} function. For
\emph{data.frame}, each first member of the cluster will be selected as
representative. In the caso of \emph{nr_list} object. The member with higher
centrality will be selected as representative.}
}
\value{
\emph{accnet} or \emph{mash} object
}
\description{
This function creates a non-redundant representation of the input object
(\emph{accnet} or \emph{mash}).
}
