% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_of_annotation.R
\name{heatmap_of_annotation}
\alias{heatmap_of_annotation}
\title{HeatMap of Annotation (VF and/or AbR)}
\usage{
heatmap_of_annotation(data, min_identity = 0.95, max_evalue = 1e-25)
}
\arguments{
\item{data}{Data frame result from annotate.}

\item{min_identity}{Minimun identity to show in figure.}

\item{max_evalue}{Maximun Evalue to show in figure.}
}
\description{
Creates a heatmap with the annotation results. User can
filter the results by identity and/or evalue. This function
uses \emph{pheatmap} instead of \emph{heatmap} funtion if
\emph{pheatmap} is installed.
}
