% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_db.R
\name{install_db}
\alias{install_db}
\title{Install PATO reference data bases}
\usage{
install_db()
}
\description{
This function download and install the reference data bases for the functions
\emph{screening} and \emph{annotation}
}
