% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnn.R
\name{knnn}
\alias{knnn}
\title{K-Nearest Neighbour Network}
\usage{
knnn(data, n_neigh, repeats = TRUE, threshold = 1)
}
\arguments{
\item{data}{An Accnet or Mash object}

\item{n_neigh}{The number of best K-Neighbours}

\item{repeats}{\emph{Boolean} Include repetitions?}

\item{threshold}{Minimum value to create an edge.}
}
\value{
Returns an \emph{igraph} object.
}
\description{
This function creates a network with the \emph{k} best neigbours
of each genome (or pangenome). The network can be buid with the
best neigbours with or without repetitions. The option \emph{repeats}
establish if the best \emph{k} neigbours includes bi-directional links
or not. Let be G_i and G_j two genomes, if G_i is one of the best
k-neighbours of G_j and G_j is one of the best k-neighbours of G_i if the
\emph{repeats} option is TRUE then G_i is removed from the k-neighbour and
substituted be the next best neighbour in other case G_i keeps in the list.
}
