% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gff_list.R
\name{load_gff_list}
\alias{load_gff_list}
\title{Load GFF3 file list}
\usage{
load_gff_list(input_files, n_cores)
}
\arguments{
\item{file}{}
}
\value{
The function returns a \code{gff_list} object the can be used as input for other functions (mmseqs, mash)
}
\description{
This function load and process a list of GFF3 files. The function extract CDS in nucleotide and
in aminoacids. It creates four folder with the GFF anotation, the FNA whole-genome sequence,
the FFN for fasta features nucleotide and the FAA with the proteome function.
}
\details{
Load GFF3 files is slower than load multiple-fasta files (nucleotide or aminoacids) so for large
datasets we recommend to load fasta files instead of GFF3.
}
