% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mash.R
\name{load_mash}
\alias{load_mash}
\title{Load MASH}
\usage{
load_mash(file, format = "matrix")
}
\arguments{
\item{file}{Path to mash output file}

\item{format}{\emph{list} or \emph{matrix} type.}
}
\value{
A \emph{mash} object
}
\description{
Load an output from MASH. The file could be the matrix distance
(-t option in mash command line) or the standart table.
}
