% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash}
\alias{mash}
\title{MASH distance estimation.}
\usage{
mash(file_list, n_cores = 4, sketch = 1000, kmer = 21,
  type = "prot")
}
\arguments{
\item{file_list}{Data frame with the full path to the genome files (gene or protein multi-fasta).}

\item{n_cores}{Number of cores to use.}

\item{sketch}{Number of sketches to use for distance estimation.}

\item{kmer}{Kmer size.}

\item{type}{Type of sequence 'nucl' (nucleotides) or 'prot' (aminoacids)}
}
\value{
A \emph{mash} object
}
\description{
MASH (Fast genome and metagenome distance estimation using MinHash)
is a fast sequence distance estimator that uses the MinHash
algorithm and is designed to work with genomes and metagenomes in the
form of assemblies or reads (https://mash.readthedocs.io/). This function
is a wrapper to execute \emph{mash} in the background and import to R as a
\emph{mash} object.
}
\note{
A \emph{mash} is a list of two element.

The first one contains a rectangular and simetric \emph{matrix} with the distances among genomes.
As a matrix has genomes as rownames and colnames

The second one is a \emph{data.table/data.frame} with all the distancies as list.
The table has the columns \emph{c("Source","Target","Dist")}
}
\references{
Mash: fast genome and metagenome distance estimation using MinHash. Ondov BD, Treangen TJ, Melsted P, Mallonee AB, Bergman NH, Koren S, Phillippy AM. Genome Biol. 2016 Jun 20;17(1):132. doi: 10.1186/s13059-016-0997-x.

Mash Screen: High-throughput sequence containment estimation for genome discovery. Ondov BD, Starrett GJ, Sappington A, Kostic A, Koren S, Buck CB, Phillippy AM. BioRxiv. 2019 Mar. doi: 10.1101/557314
}
