% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmseqs.R
\name{mmseqs}
\alias{mmseqs}
\title{MMSeqs2 Orthologous clustering.}
\usage{
mmseqs(file_list, coverage = 0.8, identity = 0.8, evalue = 1e-06,
  n_cores, cov_mode = 0, cluster_mode = 0)
}
\arguments{
\item{file_list}{Data frame with the full path to the genome files (gene or protein multi-fasta).}

\item{coverage}{Minimun coverage (length) to cluster.}

\item{identity}{Minimun Identity.}

\item{evalue}{Maximun Evalue.}

\item{n_cores}{Number of cores to use.}

\item{cov_mode}{Coverage mode:\itemize{
\item 0: Coverage of query and target
\item 1: Coverage of target
\item 2: coverage of query
\item 3: target seq. length needs be at least x% of query length
\item 4: query seq. length needs
}}

\item{cluster_mode}{Cluster mode:\itemize{
\item{0: Setcover}
\item{1: connected component}
\item{2: Greedy clustering by sequence length}
\item{3: Greedy clustering by sequence length (low mem)}
}}
}
\value{
Return a \emph{mmseq} object.
}
\description{
MMseqs2 (Many-against-Many sequence searching) is a software suite to
search and cluster huge protein and nucleotide sequence sets.
MMseqs2 is open source GPL-licensed software (https://mmseqs.com).
This function is a wrapper for the original \emph{mmseqs} executable.
The function clusters the genes or proteins in ortholog clusters and save the
representative member of each one and its annotation.
}
\note{
A \emph{mmseq} object is a list of two elements.
First contains a data.table/data.frame with four columns (Prot_genome, Prot_Prot,
Genome_genome and Genome_Prot). This is the output of MMSeqs2 and described the clustering
of the input genes/proteins. First column referes to the genome that contain the
representative gene/protein of the cluster. Second one, is the representative protein of the
cluster (i.e. the cluster name). Third colum is the genome that contains the gene/protein of
the fourth column.

In the second element we can find a data.frame/data.table with the original annotation of all
representative gene/protein of each cluster in two columns. The first one Prot_prot is
the same that the second one of the first element.
}
\references{
Steinegger M and Soeding J. MMseqs2 enables sensitive protein sequence searching for the analysis of massive data sets. Nature Biotechnology, doi: 10.1038/nbt.3988 (2017).

Steinegger M and Soeding J. Clustering huge protein sequence sets in linear time. Nature Communications, doi: 10.1038/s41467-018-04964-5 (2018).

Mirdita M, Steinegger M and Soeding J. MMseqs2 desktop and local web server app for fast, interactive sequence searches. Bioinformatics, doi: 10.1093/bioinformatics/bty1057 (2019)
}
