% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_of_annotation.R
\name{network_of_annotation}
\alias{network_of_annotation}
\title{Network of Annotation}
\usage{
network_of_annotation(data, min_identity = 0.95, max_evalue = 1e-25)
}
\arguments{
\item{data}{\emph{annotate} result.}

\item{min_identity}{Minimun identity.}

\item{max_evalue}{Max Evalue.}
}
\value{
\emph{igraph} object
}
\description{
This function builds a \emph{igraph} object with the results from \emph{annotate} function.
Allows to filter the results by identity and/or Evalue.
}
\seealso{
\code{\link{annotate}}
}
