% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_redundant.R
\name{non_redundant}
\alias{non_redundant}
\title{Non Redundant set of samples.}
\usage{
non_redundant(
  data,
  number,
  fraction,
  distance,
  tolerance = 0.05,
  max_iter = 10000,
  fast = TRUE
)
}
\arguments{
\item{data}{An \emph{accnet} or \emph{mash} object}

\item{number}{The number of non-redundant samples.}

\item{fraction}{The fraction of the whole set of non-redundant samples.}

\item{distance}{Minimun distance among samples}

\item{tolerance}{Percentage of error between the input number and the final number of samples.}

\item{max_iter}{Maximun number of search iterations.}

\item{fast}{If fast is TRUE the clustering process uses "components" in other case use "louvain"}
}
\value{
\emph{nr_list} object
}
\description{
Creates a non-redundant sub-set of samples. The function accept an
\emph{accnet} or \emph{mash} object and returns a \emph{nr_list} object.
The sub-set can be created using a certain \emph{distance} (mash o jaccard distance),
a specific \emph{number} of elements or a \emph{fraction} of the whole sample set.
The function performes an iterative seach so sometimes the exact number of returned elements
could be not the same that the specified in the input. This difference can be defined
with the \emph{threshold} parameter.
}
\seealso{
\code{\link{extract_non_redundant}}
}
