% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_redundant_hier.R
\name{non_redundant_hier}
\alias{non_redundant_hier}
\title{Non Redundant set of samples using hierarchical search.}
\usage{
non_redundant_hier(
  data,
  number,
  fraction,
  distance,
  tolerance = 0.05,
  partitions = 10,
  max_iter = 10000,
  fast = FALSE
)
}
\arguments{
\item{data}{An \emph{accnet} or \emph{mash} object.}

\item{number}{The number of non-redundant samples.}

\item{fraction}{The fraction of the whole set of non-redundant samples.}

\item{distance}{Minimun distance among samples.}

\item{tolerance}{Percentage of error between the input number and the final number of samples.}

\item{partitions}{Number of partitions to hiaerarchical search.}

\item{max_iter}{Maximun number of search iterations.}

\item{fast}{If fast is TRUE the clustering process uses "components" in other case use "fast_greddy".}
}
\value{
\emph{nr_list} object
}
\description{
Creates a non-redundant sub-set of samples. The function accept an
\emph{accnet} or \emph{mash} object and returns a \emph{nr_list} object.
The sub-set can be created using a certain \emph{distance} (mash o jaccard distance),
a specific \emph{number} of elements or a \emph{fraction} of the whole sample set.
The function performes an iterative seach so sometimes the exact number of returned elements
could be not the same that the specified in the input. This difference can be defined
with the \emph{threshold} parameter. This function perform a hierarchical search. For small
datasets (<2000) \emph{non_redundant()} could be faster. This function comsume less memory than
the original so it is better for large datasets.
}
\seealso{
\code{\link{extract_non_redundant}}
}
