% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_redundant_pangenomes.R
\name{non_redundant_pangenomes}
\alias{non_redundant_pangenomes}
\title{Non redundant pangenomes.}
\usage{
non_redundant_pangenomes(
  file_list,
  distance,
  type = "prot",
  n_cores,
  sketch = 1000,
  kmer = 21
)
}
\arguments{
\item{file_list}{A \emph{data.frame} with the path list of files or a \emph{gff_list} object.}

\item{distance}{Sequence distance (<1)}

\item{type}{"nucl" or "prot"}

\item{n_cores}{number of cores to use (default: total cores -1)}

\item{sketch}{sketch size (mash option)}

\item{kmer}{kmer size (mash option)}
}
\value{

}
\description{
Remove redundant sequences from a file list of sequences (nucleotide or protein).
}
\details{
This function has been designed to remove redundant sequences from a dataset.
Unlike other non redundant functions, this function only accept a distance threshold and has
been designed to remove very similar sequences (>99.9\%)
}
