% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Find outliers in the dataset.}
\usage{
outliers(data, threshold, n_sd, plot = TRUE)
}
\arguments{
\item{data}{An \emph{accnet} or \emph{mash} object}

\item{threshold}{threshold for outliers (outliers >= threshold)}

\item{n_sd}{Number of standard deviation as threshold.}

\item{plot}{\emph{Boolean} plot manhattan?}
}
\value{
a data.frame with the set of outliers.
}
\description{
Find outliers in the dataset using a threshold distance (mash o jaccard distance) or by
standard deviation times. This function also plot a manhattan plot to visualize the
structure of the distance distribution of the input object.
}
\seealso{
\code{\link{remove_outliers}}
}
