% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenomes_from_files.R
\name{pangenomes_from_files}
\alias{pangenomes_from_files}
\title{Creates Pangenomes from file list}
\usage{
pangenomes_from_files(
  files,
  min_pange_size = 10,
  min_prot_freq = 2,
  file_type = "prot",
  distance,
  cluster,
  coverage = 0.8,
  identity = 0.8,
  evalue = 1e-06,
  n_cores,
  cov_mode = 0,
  cluster_mode = 0
)
}
\arguments{
\item{files}{A data.frame of one column whit the path of the files.}

\item{min_pange_size}{Minimun number of genomes per pan-genome}

\item{min_prot_freq}{Mininum frequency of a protein (minimun number of pan-genomes in where is present)}

\item{file_type}{Type of fasta file (prot or nucl)}

\item{distance}{Maximun distance (mash distance) among genomes in each pangenome. (exlude \emph{cluster})}

\item{cluster}{(optional, exclude distance). Data.frame of two columns (file, cluster) describing the clutering of the files}

\item{coverage}{Minimun coverage (length) to cluster.}

\item{identity}{Minimun Identity.}

\item{evalue}{Maximun Evalue.}

\item{n_cores}{Number of cores to use.}

\item{cov_mode}{Coverage mode:\itemize{
\item 0: Coverage of query and target
\item 1: Coverage of target
\item 2: coverage of query
\item 3: target seq. length needs be at least x\% of query length
\item 4: query seq. length needs
}}

\item{cluster_mode}{Cluster mode:\itemize{
\item{0: Setcover}
\item{1: connected component}
\item{2: Greedy clustering by sequence length}
\item{3: Greedy clustering by sequence length (low mem)}
}}
}
\value{
An \emph{accnet} with an extra membership table.
}
\description{
This function perform a clustering using a designated maximun distance among samples
and creates a set of pangenomes using it.
Pan-genomes can be filtered by genome frequency (i.e. the number of genomes that belong to
the pangenome). \emph{pangenomes_from_list()} builds a new \emph{accnet} object that relates
the pangenome with the genes/proteins. The user can filter the proteins that are included
by the presence frequency in the set.
}
