% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenomes_from_files_mmseqs.R
\name{pangenomes_from_files_mmseqs}
\alias{pangenomes_from_files_mmseqs}
\title{Individual pangenome clustering from file-list}
\usage{
pangenomes_from_files_mmseqs(
  files,
  i,
  coverage,
  identity,
  evalue,
  n_cores,
  cov_mode,
  cluster_mode
)
}
\arguments{
\item{files}{a list of files (genomes of the pangenome)}

\item{i}{The pangenome number}

\item{coverage}{min coverage  for homologous cluster}

\item{identity}{min identitity for homologous cluster}

\item{evalue}{max E-value for homologous cluster}

\item{n_cores}{number of cores to use}

\item{cov_mode}{coverage mode}

\item{cluster_mode}{cluster mode}
}
\value{
A \emph{list} with two tables, the membership of the pangenome,
and the gene/protein frequency.
}
\description{
This is an internal function for the workflow of pangenomes
}
