% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagenomes_from_mmseqs.R
\name{pangenomes_from_mmseqs}
\alias{pangenomes_from_mmseqs}
\title{Creates Pangenomes from mmseqs object}
\usage{
pangenomes_from_mmseqs(
  data,
  cluster,
  min_freq = 0,
  max_freq = 1,
  min_pangenome_size = 1
)
}
\arguments{
\item{data}{A mmseq object}

\item{cluster}{A data.frame with two columns: strain label (ID) and cluster.
Colnames are not required. Cluster column can be any categorical variable.}

\item{min_freq}{Minimun gene/protein frequency to include in the pangenomes.}

\item{max_freq}{Maximun gene/protein frequency to include in the pangenomes.}

\item{min_pangenome_size}{Minimun number of genomes to create a new pangenome.}
}
\value{
An \emph{accnet} object.
}
\description{
Creates clusters of genomes (pan-genome) using a list of pre-defined clusters.
The list of cluster can be from external source (e.g. mlst or BAPS) or from an
internal function (e.g \emph{clustering} function). Moreover, pan-genomes can be
filtered by genome frequency (i.e. the number of genomes that belong to
the pangenome). pangenomes builds a new \emph{accnet} object that relates the pangenome
with the genes/proteins. The user can filter the proteins that are included by
the presence frequency.
}
\seealso{
\code{\link{clustering}}
}
