% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmidome.R
\name{plasmidome}
\alias{plasmidome}
\title{Plasmidome: Extract the plasmid sequences from a set of genomes}
\usage{
plasmidome(gff_list, specie)
}
\arguments{
\item{gff_list}{A gff_list object (see \code{load_gff_list()})}

\item{specie}{mlplasmid accept as species Enterococcus faecium, Escherichia coli
or Klebsiella pneumoniae}
}
\value{

}
\description{
This function uses \emph{mlplasmids} to find the contigs belong to plasmids.
The package \emph{mlplasmids} must be installed in your computer (see datails).
Nowadays \emph{mlplasmdis} only process Enterococcus faecium, Escherichia coli
or Klebsiella pneumoniae genomes
}
\details{
The input genomes must be in \emph{gff_list} format (see \emph{load_gff_list()} function)
Plasmidome find the contigs that belong to plasmids and creates a new GFF structure
with all the features information (FAA, FNA, FFN and GFF).

By default PATO does not install \emph{mlplasmids} because it is deposits in
its own gitlab repository (https://gitlab.com/sirarredondo/mlplasmids).

The authors define \emph{mlplasmids} as:

\cite{mlplasmids consists of binary classifiers to predict contigs either as
plasmid-derived or chromosome-derived. It currently classifies short-read
contigs as chromosomes and plasmids for Enterococcus faecium, Escherichia coli
and Klebsiella pneumoniae. Further species will be added in the future. The
provided classifiers are Support Vector Machines which were previously
optimized and selected versus other existing machine-learning techniques
(Logistic regression, Random Forest..., see also Reference). The classifiers
use pentamer frequencies (n = 1,024) to infer whether a particular contig
is plasmid- or chromosome- derived.}

To install \emph{mlplasmids} you must have installed \emph{devtools} (usually
you have because PATO requires that you have devtools installed too). Then type:

\code{devtools::install_git("https://gitlab.com/sirarredondo/mlplasmids")}
}
\examples{


}
\references{
Arredondo-Alonso et al., Microbial Genomics 2018;4 DOI 10.1099/mgen.0.000224
}
