% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove the outliers of a data set (mash or accnet)}
\usage{
remove_outliers(data, outliers)
}
\arguments{
\item{data}{An \emph{accnet} or \emph{mash} object}

\item{outliers}{A one data frame with a column "Source".
It must contains the elements to remove}
}
\value{
An \emph{accnet} or \emph{mash} object without the removed elements.
}
\description{
Remove the outliers (or any desire element) of an \emph{accnet} or \emph{mash} object.
}
\seealso{
\code{\link{outliers}}
}
