% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{screening}
\alias{screening}
\title{Screening}
\usage{
screening(
  data,
  type = "nucl",
  database = c("AbR", "VF_A", "VF_B"),
  query = "all"
)
}
\arguments{
\item{data}{A \code{mmseq} object}

\item{type}{user must be specified if the data set is nucleotide, wgs or protein.}

\item{database}{A vector with the query databases:
\itemize{
\item \emph{AbR}: Antibiotic resistance database (ResFinder)
\item \emph{VF_A} VFDB core dataset (genes associated with experimentally verified VFs only)
\item \emph{VF_B} VFDB full dataset (all genes related to known and predicted VFs)
\item \emph{bacmet} BacMet dataset (genes associated with Biocide and Metal resistance)
}}

\item{query}{"all" or "accessory". It perform the DB from whole
protein dataset or just from the accessory}
}
\value{

}
\description{
Screening
}
