% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedness.R
\name{sharedness}
\alias{sharedness}
\title{Creates a matrix of sharedness}
\usage{
sharedness(data)
}
\arguments{
\item{data}{An \emph{accnet} object.}
}
\value{
A matrix with the count of the genes/proteins shared between
each pair of samples.
}
\description{
Count the number of genes/proteins that share the samples among them.
}
