% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_network.R
\name{similarity_network}
\alias{similarity_network}
\title{Similarity Network}
\usage{
similarity_network(data, threshold)
}
\arguments{
\item{data}{An object of class \emph{accnet} or \emph{mash}}

\item{threshold}{Maximun distance to include in the network.}
}
\value{
An \emph{igraph} object.
}
\description{
The function creates a similarity network connecting genomes
with 'distances' lower than the especified threshold.
}
\details{
The distance must to be addecuated for the type of data.
MASH values rangin between 0-0.05 intra-specie mean while accnet distances
could be higher.
}
