% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singles.R
\name{singles}
\alias{singles}
\title{Find singles (genes/proteins) in a AccNET object}
\usage{
singles(data, cluster)
}
\arguments{
\item{data}{An \emph{accnet} object.}

\item{cluster}{A data.frame with two columns. The first one must match with
\emph{Source} column of Accnet object (usually the genomes/samples). Second one
must contain a categorical variable (cluster, group, ST, origin, source ...)}
}
\value{
A data.frame with Genome/cluster, the genes/proteins belong to
Genome/Cluster and its annotation.
}
\description{
\emph{singles()} return the genes/proteins that are only present in each
genome or cluster of genomes.
}
