% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_map.R
\name{snps_map}
\alias{snps_map}
\title{SNPs Map}
\usage{
snps_map(core_snp_genome)
}
\arguments{
\item{core_snp_genome}{A \emph{core_snp_genome} object.}
}
\value{
A \emph{ggplot2} object
}
\description{
snps_map is a function that makes a similar plot to manhattan plots.
This function plot the cumulative number of snps per each contig-position pair.
}
