% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_pairwaise.R
\name{snps_pairwaise}
\alias{snps_pairwaise}
\title{Direct SNPs pairwaise.}
\usage{
snps_pairwaise(file_list, type, n_cores)
}
\arguments{
\item{n_cores}{}
}
\value{

}
\description{
This function calculates All-vs-All SNPs number. At different of \emph{core_snps_matrix} this function
calculates the raw number of SNPs among each pair of sequences avoiding the bias of the core genome size and/or
any reference.
}
\examples{

}
