% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twins.R
\name{twins}
\alias{twins}
\title{Finds Twins into AccNET object}
\usage{
twins(data)
}
\arguments{
\item{data}{An \emph{accnet} object}
}
\value{
A data.frame() with columns "Target", "Twin" and "Size".
\itemize{
\item Target:  Column with gene/protein ID.
\item Twin:  Twin label
\item Size:  Twin group size (number of genes/proteins size)
}
}
\description{
Twins nodes are those nodes that share exactly the same edges. In the case
of an \emph{accnet} object twin are those genes/proteins that you can find
in the exactly the same set of genomes/samples.
}
