% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{umap_plot}
\alias{umap_plot}
\title{Plot UMAP representation of a data set.}
\usage{
umap_plot(data, plot = TRUE, cluster, ...)
}
\arguments{
\item{data}{An \emph{accnet} or \emph{mash} object}

\item{plot}{print plot?}

\item{cluster}{Two column data.frame with the nodes (column 1) and the cluster
or classification(column 2)}

\item{...}{Passed to uwot::umap function.}
}
\value{
A matrix of optimized coordinates.
}
\description{
UMAP (Uniform Manifold Approximation and Projection for Dimension Reduction) is a dimension
reduction technique that can be used for visualisation similarly to t-SNE, but also for
general non-linear dimension reduction. This function uses UMAP to represent the population structure
of the data set (\emph{accnet} or \emph{mash})
}
\references{
McInnes, L, Healy, J, UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction, ArXiv e-prints 1802.03426, 2018

James Melville (2019). uwot: The Uniform Manifold Approximation and Projection (UMAP) Method for
Dimensionality Reduction. R package version 0.1.4. https://CRAN.R-project.org/package=uwot
}
\seealso{
\code{\link{uwot}}
}
