% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_genome.R
\name{core_genome}
\alias{core_genome}
\title{Core-Genome Alignment}
\usage{
core_genome(data, type, n_cores)
}
\arguments{
\item{data}{An \emph{mmseqs} object}

\item{type}{Type of sequence 'nucl' or 'prot'}
}
\value{
A core_genome object (a data.frame with two columns: fasta header and sequence)
}
\description{
Find and creates a core-genome alignment. Unlike \emph{core_plots()}
this function find the hard core-genome (genes presence in 100\% of genomes
and without repetitions). The function takes a \emph{mmseqs()} output, so the deffinition of
the paralogous genes of the core-genome (similarity, coverage and/or e-value)
depends on the \emph{mmseqs()} parameters.
}
\details{
The function perform a pseudo-msa per each paralog using the function
\emph{result2msa} of \emph{mmseqs2}.This approach is much faster than
classical MSA (clutal, mafft or muscle) but is less accurate. Taking into account
that most of the phylogenetic inference software only takes variant columns with
no insertions or deletion, there are not to many difference in the final phylogenetic trees.

\emph{core_genome()} can build a core-genome alingment of thusands of genomes in minutes.
}
