% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_redundant_pangenomes.R
\name{non_redundant_pangenomes}
\alias{non_redundant_pangenomes}
\title{Title}
\usage{
non_redundant_pangenomes(
  files,
  distance,
  type = "prot",
  n_cores,
  sketch = 1000,
  kmer = 21
)
}
\value{

}
\description{
Title
}
